fix error

+++ xturing/engines/llama_utils/llama.py
@@ -993,9 +993,9 @@ class LlavaLlamaModel(LlamaModel):
                     select_hidden_state = image_forward_outs.hidden_states[select_hidden_state_layer]
                     image_features = select_hidden_state[:, 1:]
             if type(images) is list:
-                image_features = [self.mm_projector(image_feature)[0] for image_feature in image_features]
+                image_features = [self.mm_projector(image_feature.to(torch.float16))[0] for image_feature in image_features]
             else:
-                image_features = self.mm_projector(image_features)
+                image_features = self.mm_projector(image_features.to(torch.float16))
             dummy_image_features = torch.zeros(256, 1024, device=inputs_embeds.device, dtype=inputs_embeds.dtype)
             dummy_image_features = self.mm_projector(dummy_image_features)
